完善movement，添加targetMovement,完善util库

+++ src/utils/vec2.h
@@ -32,8 +32,8 @@ public:
 	T operator*(const Vec2<T>&other)const;
 
 	void  Set(T xx, T yy);
-	float GetLength()const;
-	float GetLengthSquared()const;
+	double GetLength()const;
+	double GetLengthSquared()const;
 
 	Vec2<T> Normalized();
